Identification of a material by its spectrum
The light reflected by an object to the space and absorbed by the imaging system of a satellite have a peculiar spectral footprint for every material and sunlight exposition condition. With theses programs it's analyse the spectral response of the satellite combined to a given light to know the optical coefficients that the satellite will send back to the earth as a digital image in "coefficient solver v1", and later it's compared to a database of known material in "finder_metric_v1". These proposal it's in two stages to improve the speed of the system and avoid repetition of calculations, that's because the amount of expected information is huge.